class diagram update

+++ C#/SEP3/FileData/DAOs/OrderFileDao.cs
@@ -15,7 +15,7 @@ public class OrderFileDao : IOrderDao
 
     public Task<Order> CreateAsync(Order order)
     {
-        int orderId = 1;
+        long orderId = 1;
         if (context.Orders.Any())
         {
             orderId = context.Orders.Max(c => c.Id);
